Add more formulas to calculate distance : - Use double instead of ints since values are between 0 and 1

+++ src/irdm/descriptors/Descriptor.java
@@ -9,7 +9,7 @@ public interface Descriptor {
 
     }
 
-    public int distance(Descriptor o2);
+    public double distance(Descriptor o2);
 
     public static double distance(Descriptor o1, Descriptor o2) {
         return o1.distance(o2);
